refactor(mcp): migrate agent_router server to new pattern
